adjusted paths and deleted folders for AWS

+++ L1_Emo_PPI.sh
@@ -6,14 +6,18 @@ subj=$2
 
 basedir=`pwd`
 cd ..
-MAINDATADIR=`pwd`/data
-MAINOUTPUTDIR=`pwd`/outputs
+MAINDATADIR=/s3/hcp
+MAINOUTPUTDIR=`pwd`/Analysis
 cd $basedir
 
 OUTPUT=${MAINOUTPUTDIR}/${subj}/MNINonLinear/Results/tfMRI_${task}_${run}/L1_Emotion_PPI
 DATA=${MAINOUTPUTDIR}/${subj}/MNINonLinear/Results/tfMRI_${task}_${run}/L1_Emotion_Act.feat/filtered_func_data.nii.gz
 NVOLUMES=`fslnvols ${DATA}`
-rm -rf ${OUTPUT}.feat
+
+#remove if output file exists
+if [ -e ${OUTPUT}.feat ]; then
+  rm -rf ${OUTPUT}.feat
+fi
 
 #EV files
 EVFEAR=${MAINDATADIR}/${subj}/MNINonLinear/Results/tfMRI_${task}_${run}/EVs/fear.txt
@@ -25,7 +29,7 @@ TIMECOURSE=${MAINOUTPUTDIR}/${subj}/MNINonLinear/Results/tfMRI_${task}_${run}/my
 fslmeants -i $DATA -o $TIMECOURSE -m $MASK
 
 #find and replace: run feat for smoothing
-ITEMPLATE=${basedir}/templates/L1_EMO_PPI.fsf
+ITEMPLATE=${basedir}/templates/L1_Emo_PPI.fsf
 OTEMPLATE=${MAINOUTPUTDIR}/${subj}/MNINonLinear/Results/tfMRI_${task}_${run}/L1_Emotion_PPI.fsf
 sed -e 's@OUTPUT@'$OUTPUT'@g' \
 -e 's@DATA@'$DATA'@g' \
@@ -37,7 +41,7 @@ sed -e 's@OUTPUT@'$OUTPUT'@g' \
 feat $OTEMPLATE
 
 #delete unused files
-#rm -rf ${OUTPUT}.feat/filtered_func_data.nii.gz
+rm -rf ${OUTPUT}.feat/filtered_func_data.nii.gz
 rm -rf ${OUTPUT}.feat/stats/res4d.nii.gz
 rm -rf ${OUTPUT}.feat/stats/corrections.nii.gz
 rm -rf ${OUTPUT}.feat/stats/threshac1.nii.gz
